requirements.txt
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
Werkzeug==3.0.1
python-dotenv==1.0.0
pymongo==4.6.0
Pillow==10.1.0
datetime==5.3
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
mlxtend==0.22.0
mistralai==0.0.10
